# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the administration API for the ImLate tracking system.",
        "title": "ImLate Admin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin-api/reports/visits": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated visitor attendance information aggregated per visitor per day for a given date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-reports"
                ],
                "summary": "Get visitor attendance report",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, inclusive)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, inclusive)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by student flag",
                        "name": "is_student",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year group",
                        "name": "year_group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sign status (not_signed, signed_in, signed_out)",
                        "name": "sign_status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Records per page (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ReportsVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated, field-selectable visitor attendance report with multi-value filters and server-side sorting.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-reports"
                ],
                "summary": "Search visitor attendance report",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.PostReportsVisitsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.PostReportsVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin-api/track/visit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Manually records a visitor sign-in/sign-out event on behalf of a visitor by an admin.\nUsed when a visitor forgot their key, lost it, or is otherwise unable to use it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-tracking"
                ],
                "summary": "Manual visitor track",
                "parameters": [
                    {
                        "description": "Track request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ManualTrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ManualTrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_buzyka_imlate_internal_domain_entity.Visitor": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "grade": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "is_student": {
                    "type": "boolean"
                },
                "isams_divisions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "isams_id": {
                    "type": "integer"
                },
                "isams_school_id": {
                    "type": "string"
                },
                "isams_year_group_id": {
                    "type": "integer"
                },
                "keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_buzyka_imlate_internal_usecase_adminapi.PostReportsVisitsFilters": {
            "type": "object",
            "properties": {
                "is_student": {
                    "type": "boolean"
                },
                "sign_status": {
                    "description": "SignStatus filters rows by sign status. Allowed values: signed_in, signed_out, not_signed.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "year_group": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_buzyka_imlate_internal_usecase_adminapi.PostReportsVisitsOrder": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "field": {
                    "type": "string",
                    "enum": [
                        "sign_status",
                        "year_group",
                        "name",
                        "surname",
                        "visit_date",
                        "visits_count"
                    ]
                }
            }
        },
        "github_com_buzyka_imlate_internal_usecase_adminapi.ReportsVisitsResponseItem": {
            "type": "object",
            "properties": {
                "duration_minutes": {
                    "type": "integer"
                },
                "is_student": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sign_status": {
                    "type": "string"
                },
                "signed_in": {
                    "type": "string"
                },
                "signed_out": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "visit_date": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "integer"
                },
                "visits_count": {
                    "type": "integer"
                },
                "year_group": {
                    "type": "integer"
                }
            }
        },
...
        "internal_http_controller_adminapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_adminapi.ManualTrackRequest": {
            "type": "object",
            "required": [
                "visitor_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "signed_in": {
                    "type": "boolean"
                },
                "visitor_id": {
                    "type": "integer"
                }
            }
        },
        "internal_http_controller_adminapi.ManualTrackResponse": {
            "type": "object",
            "properties": {
                "track_date": {
                    "type": "string"
                },
                "track_type": {
                    "type": "string"
                },
                "visitor": {
                    "$ref": "#/definitions/github_com_buzyka_imlate_internal_domain_entity.Visitor"
                }
            }
        },
...
        "internal_http_controller_adminapi.PostReportsVisitsRequest": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "Fields selects which optional fields to include in each row.\nComputed fields (visit_date, sign_status, visits_count, signed_in, signed_out, duration_minutes)\nare always present regardless of this list.\nAllowed values: visitor_id, visit_date, name, surname, is_student, year_group,\nvisits_count, sign_status, signed_in, signed_out, duration_minutes, email, image.\nOmit or leave empty to return all optional fields.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filters": {
                    "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_adminapi.PostReportsVisitsFilters"
                },
                "from": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "order": {
                    "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_adminapi.PostReportsVisitsOrder"
                },
                "page": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_adminapi.PostReportsVisitsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_http_controller_adminapi.ReportsVisitsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_adminapi.ReportsVisitsResponseItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT token in format: \"Bearer \u003cJWT\u003e\".\nToken lookup order: Authorization header, query parameter `token`, cookie `jwt`.\nProtected admin routes require admin role.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}